# docs/research/main_specs/1.0.0-mvp/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Helix OTA Control Plane API
  version: 1.0.0-MVP
  summary: REST API for the Helix OTA modular-monolith control plane (Android 15 first).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  description: >
    Control-plane endpoints (auth, devices, artifacts, releases, deployments, client
    update-check + telemetry). JSON responses negotiate Brotli/gzip; the referenced
    artifact-byte path is served identity-encoded, ZIP_STORED, with HTTP Range support.
    Response-header conventions (applied to every negotiated JSON response, not repeated per
    operation): 'Vary: Accept-Encoding' is always set (the Vary header component), and 'X-Request-Id'
    is always returned. Rate-limited routes additionally return RateLimit-Limit / RateLimit-Remaining
    / RateLimit-Reset (draft IETF RateLimit fields); on 429 a Retry-After header is returned.
  contact:
    name: Helix OTA control-plane team
  # license: intentionally omitted — project license is UNVERIFIED in this revision.
servers:
  - url: https://{host}/api/v1
    description: Helix OTA control plane (HTTP/3 primary, HTTP/2 fallback).
    variables:
      host:
        default: helix.example
...
security:
  - bearerAuth: []
paths:
...
  /artifacts/{artifactId}:
    get:
      tags: [artifacts]
      operationId: getArtifact
      summary: Artifact metadata (not the bytes; bytes are referenced from the update-check).
      parameters:
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Artifact metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artifact' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2/JWT access token (admin/operator/viewer) or device-scoped token (role=device).
        RBAC enforced per-route; device tokens may act only on their own deviceId.
    provisioningToken:
      type: http
      scheme: bearer
      description: >
        Alternative auth for POST /devices/register only: a provisioning token that authorizes a
        device to enrol itself when no operator bearer is available. The global bearerAuth (operator
        role) is the primary path; this scheme is the alternative (either satisfies the operation).
...
  parameters:
...
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthenticated:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but lacks the required role or resource ownership.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    SemVer:
      type: string
      pattern: '^\d+\.\d+\.\d+$'
      examples: ['1.1.0']
    Os:
      type: string
      enum: [android]
      description: MVP targets Android 15 first; Linux/universal is a later phase.
    Sha256Hex:
      type: string
      pattern: '^[a-f0-9]{64}$'
      description: Lowercase-hex SHA-256.
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_FAILED
                - CONFLICT
                - UNSUPPORTED_MEDIA_TYPE
                - PAYLOAD_TOO_LARGE
                - RATE_LIMITED
                - SIGNATURE_INVALID
                - HASH_MISMATCH
                - VERSION_NOT_MONOTONIC
                - INTERNAL
            message: { type: string }
            request_id: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                properties:
                  field: { type: string }
                  issue: { type: string }
...
    Artifact:
      type: object
      required: [artifact_id, sha256, size, os, target_model, version, verified, uploaded_at]
      additionalProperties: false
      properties:
        artifact_id: { type: string, format: uuid }
        sha256: { $ref: '#/components/schemas/Sha256Hex' }
        size: { type: integer, format: int64 }
        os: { $ref: '#/components/schemas/Os' }
        target_model: { type: string }
        version: { $ref: '#/components/schemas/SemVer' }
        storage_ref: { type: string, description: Opaque object-store reference (MinIO/S3). }
        verified: { type: boolean }
        uploaded_at: { type: string, format: date-time }
...